requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
Pillow>=10.0
psycopg2-binary>=2.9
python-decouple>=3.8
